Feature - Character 체력 Prefabs로 제작 중

+++ Assets/ScriptableObject/Scripts/DefaultStatsData.cs
@@ -8,7 +8,7 @@ public class DefaultStatsData : ScriptableObject
 {
     [Header("Stats Info")]
     public int level;
-    public float max_health;
+    public int max_health;
     public int coin;
     public float speed;
 }
